requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
httpx==0.28.1
pytest==8.4.1
langgraph==0.6.6
anthropic==0.67.0
streamlit==1.49.1
python-dotenv==1.1.0
